Remove file device_elementwise_extension.hpp and move NormalizeInInfer operation to element_wise_operation.hpp
